api/pyproject.toml
[tool.poetry]
name = "stocks"
version = "0.1.0"
description = ""
authors = ["John Anderson <sontek@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
uvicorn = "^0.13.3"
fastapi = "^0.63.0"
python-dotenv = "^0.15.0"
gunicorn = "^20.0.4"
uvloop = "^0.14.0"
httptools = "^0.1.1"
python-multipart = "^0.0.5"
virtualenv = "^20.4.2"
appdirs = "^1.4.4"
redis = "^3.5.3"
hiredis = "^1.1.0"
requests = "^2.25.1"
celery = {extras = ["redis"], version = "^5.0.5"}
python-dateutil = "^2.8.1"
pytz = "^2021.1"
orjson = "^3.4.8"
pandas = "^1.2.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
web-pdb = "^1.5.6"
black = "^20.8b1"
flake8 = "^3.8.4"
pytest-docker = "^0.10.1"
mock = "^4.0.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
